Agrego validacion a las operaciones de la API
Las operaciones que se realizan contra el app server ahora
son validadas, en base al token que debe enviar el usuario.
Por ahora el unico token valido es 'token' hasta que
se implemente desde el lado del shared server el servicio
de verificacion.

+++ src/ServerController.cpp
@@ -14,12 +14,14 @@ using namespace Mongoose;
 
 ServerController::ServerController(Logger* logger){
     this->dbhandler = new DBhandler();
+    this->validator = new Validator();
     this->logger = logger;
 
 }
 
 ServerController::~ServerController(){
     delete this->dbhandler;
+    delete this->validator;
 }
 
 void ServerController::setup(){
@@ -39,32 +41,49 @@ void ServerController::setup(){
 
 
 void ServerController::hellojson(Mongoose::Request &request, Mongoose::JsonResponse &response){
-    response["data"] = "Testing Json response";
-    logger->log("Testing OK", Log_type::INFO);
+    if ( validator->validate_token( request.get("api_token", "") ) ){
+        response["data"] = "Testing Json response";
+        logger->log("Testing OK", Log_type::INFO);     
+    } else {
+        response["data"] = "El token recibido no es valido";
+        logger->log("El token recibido no es valido", Log_type::WARN);
+    }
+   
 };
 
 void ServerController::get_song(Mongoose::Request &request, Mongoose::JsonResponse &response){
-
-    int id_song = stoi(request.get("id_song", ""));
-    string song_file = dbhandler->get(id_song);
-    response["song_file"] = song_file;
-
-    logger->log("Se envio la cancion: " + song_file, Log_type::INFO);
+    if ( validator->validate_token( request.get("api_token", "") ) ){
+        int id_song = stoi(request.get("id_song", ""));
+        string song_file = dbhandler->get(id_song);
+        response["song_file"] = song_file;
+        logger->log("Se envio la cancion: " + song_file, Log_type::INFO);
+    } else {
+        response["data"] = "El token recibido no es valido";
+        logger->log("El token recibido no es valido", Log_type::WARN);
+    }
 };
 
 void ServerController::add_song(Mongoose::Request &request, Mongoose::JsonResponse &response){
-    int id_song = stoi(request.get("id_song", ""));
-    string song_file = request.get("song_file", "");
-    dbhandler->insert(id_song, song_file);
-
-    response["data"] = "Song added !";
-    logger->log("Se agrego la cancion: " + song_file + " a la base de datos", Log_type::INFO);
+    if ( validator->validate_token( request.get("api_token", "") ) ){
+        int id_song = stoi(request.get("id_song", ""));
+        string song_file = request.get("song_file", "");
+        dbhandler->insert(id_song, song_file);
+        response["data"] = "Song added !";
+        logger->log("Se agrego la cancion: " + song_file + " a la base de datos", Log_type::INFO);
+    } else {
+        response["data"] = "El token recibido no es valido";
+        logger->log("El token recibido no es valido", Log_type::WARN);
+    }
 };
 
 void ServerController::del_song(Mongoose::Request &request, Mongoose::JsonResponse &response){
-    int id_song = stoi(request.get("id_song", ""));
-    dbhandler->deleteByKey(id_song);
-
-    response["data"] = "Song deleted !";
-    logger->log("Se borro la cancion: " + to_string(id_song) + " de la base de datos", Log_type::INFO);
+    if ( validator->validate_token( request.get("api_token", "") ) ){
+        int id_song = stoi(request.get("id_song", ""));
+        dbhandler->deleteByKey(id_song);
+        response["data"] = "Song deleted !";
+        logger->log("Se borro la cancion: " + to_string(id_song) + " de la base de datos", Log_type::INFO);
+    } else {
+        response["data"] = "El token recibido no es valido";
+        logger->log("El token recibido no es valido", Log_type::WARN);
+    }
 };
